requirements.txt
fastapi
uvicorn[standard]
numpy==1.26.4
pandas==2.2.2
scipy==1.13.0
torch==2.3.0
transformers==4.41.2
gradio==4.31.4
matplotlib==3.9.0
seaborn==0.13.2
deepgram-sdk==2.12.0